pyproject.toml
[project]
name = "hex-mcp"
version = "0.1.6"
description = "A MCP server for Hex"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
 "backoff>=2.2.1",
 "mcp>=1.3.0",
 "python-dotenv>=1.0.1",
 "httpx>=0.27.0",
 "typer>=0.15.2",
 "pyyaml>=6.0.2",
]
[[project.authors]]
name = "Franccesco Orozco"
email = "franccesco@codingdose.info"

[build-system]
requires = [ "hatchling",]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "mcp[cli]>=1.3.0",
]

[project.scripts]
hex-mcp = "hex_mcp:main_entry"

[project.optional-dependencies]
dev = [
  "mcp[cli]>=1.3.0",
]
